WIP: reformatting queue card logic to be more readable
