+++ client/src/components/home/index.tsx
@@ -33,7 +33,7 @@ import parseMentor, {
   defaultMentorInfo,
 } from "components/my-mentor-card/mentor-info";
 import NavBar from "components/nav-bar";
-import { isAnswerComplete, launchMentor } from "helpers";
+import { launchMentor } from "helpers";
 import {
   QuestionEdits,
   useWithReviewAnswerState,
@@ -43,16 +43,14 @@ import { useWithTraining } from "hooks/task/use-with-train";
 import withAuthorizationOnly from "hooks/wrap-with-authorization-only";
 import useActiveMentor from "store/slices/mentor/useActiveMentor";
 import { useMentorEdits } from "store/slices/mentor/useMentorEdits";
-import { User, Subject, UserRole, Status, Answer } from "types";
+import { User, Subject, UserRole, Status } from "types";
 import withLocation from "wrap-with-location";
 import RecordingBlockItem from "./recording-block";
 import { useWithRecordState } from "hooks/graphql/use-with-record-state";
 import UploadingWidget from "components/record/uploading-widget";
-import QueueBlock from "./queue-block";
-import { fetchMentorRecordQueue, setRecordQueueGQL } from "api";
-import useQuestions from "store/slices/questions/useQuestions";
+import RecordQueueBlock from "./record-queue-block";
 import { useWithLogin } from "store/slices/login/useWithLogin";
-import { QuestionState } from "store/slices/questions";
+import { useWithRecordQueue } from "hooks/graphql/use-with-record-queue";
 
 const ColorTooltip = withStyles({
   tooltip: {
@@ -162,14 +160,8 @@ function HomePage(props: {
   const mentorInfo = getData((ms) =>
     ms.data ? parseMentor(ms.data) : defaultMentorInfo
   );
-  const mentorAnswers: Answer[] = getData((state) => state.data?.answers);
-
-  const mentorQuestions = useQuestions(
-    (state) => state.questions,
-    mentorAnswers?.map((a) => a.question)
-  );
   const recordState = useWithRecordState(props.accessToken, props.search);
-
+  const { recordQueue } = useWithRecordQueue(props.accessToken);
   const [recordSubjectTooltipOpen, setRecordSubjectTooltipOpen] =
     useState<boolean>(false);
   const [buildTooltipOpen, setBuildTooltipOpen] = useState<boolean>(false);
@@ -186,7 +178,6 @@ function HomePage(props: {
   const [localHasSeenTooltips, setLocalHasSeenTooltips] = useState(false);
   const { userSawSplashScreen, userSawTooltips } = loginState;
   const [initialLoadComplete, setInitialLoadComplete] = useState(false);
-  const [answerStatuses, setAnswerStatuses] = useState<Status[]>([]);
 
   const hasSeenSplash = Boolean(
     loginState.state.user?.firstTimeTracking.myMentorSplash ||
@@ -197,48 +188,6 @@ function HomePage(props: {
       loginState.state.user?.firstTimeTracking.tooltips ||
       localHasSeenTooltips
   );
-
-  // TODO: Move all QueueBlock logic/data to a hook
-  const [queueIDList, setQueueIDList] = useState<string[]>([]);
-
-  useEffect(() => {
-    fetchMentorRecordQueue(props.accessToken).then((queueIDList) => {
-      setQueueIDList(queueIDList);
-    });
-  }, []);
-
-  useEffect(() => {
-    if (!mentorAnswers) {
-      return;
-    }
-    if (mentorAnswers.length && !answerStatuses.length) {
-      setAnswerStatuses(mentorAnswers.map((answer) => answer.status));
-      return;
-    }
-    const differentLength = mentorAnswers.length !== answerStatuses.length;
-    let differentStatus = false;
-    if (!differentLength) {
-      mentorAnswers.forEach((answer, i) => {
-        if (answerStatuses[i] !== answer.status) {
-          differentStatus = true;
-        }
-      });
-    }
-    if (differentStatus || differentLength) {
-      setAnswerStatuses(mentorAnswers.map((answer) => answer.status));
-    }
-  }, [mentorAnswers]);
-
-  useEffect(() => {
-    if (!mentorAnswers || queueIDList.length == 0) {
-      return;
-    }
-    const idListAfterRemoval = removeCompleteAnswerFromQueue(
-      queueIDList,
-      mentorAnswers
-    );
-    setQueueIDList(idListAfterRemoval);
-  }, [answerStatuses]);
 
   useEffect(() => {
     const _blocks = reviewAnswerState.getBlocks();
@@ -366,36 +315,6 @@ function HomePage(props: {
         question.originalQuestion.clientId || question.originalQuestion._id
       );
     }
-  }
-  function removeCompleteAnswerFromQueue(
-    queueIDList: string[],
-    mentorAnswers: Answer[]
-  ): string[] {
-    // remove answered questions from queue
-    const mentorAnswersInRecordQueue = mentorAnswers.filter((mentorAnswer) =>
-      queueIDList.includes(mentorAnswer.question)
-    );
-    const newQueueIdList = queueIDList.filter((questionId) => {
-      const answer = mentorAnswersInRecordQueue.find(
-        (a) => a.question === questionId
-      );
-      if (!answer) {
-        return false;
-      }
-      return !isAnswerComplete(answer, undefined, mentorType);
-    });
-    setRecordQueueGQL(props.accessToken, newQueueIdList);
-    return newQueueIdList;
-  }
-
-  // get question string
-  function getQueueQuestions(
-    queueIDList: string[],
-    mentorQuestions: Record<string, QuestionState>
-  ) {
-    return queueIDList.map(
-      (id) => mentorQuestions[id]?.question?.question || ""
-    );
   }
 
   function closeDialog() {
@@ -538,22 +457,19 @@ function HomePage(props: {
           </Select>
         </ColorTooltip>
       </div>
-      {queueIDList.length != 0 ? (
-        <List
-          data-cy="queue-block"
+      {recordQueue.length != 0 ? (
+        <ListItem
           style={{
             flex: "auto",
             backgroundColor: "#eee",
           }}
         >
-          <ListItem>
-            <QueueBlock
-              classes={classes}
-              queueIDList={queueIDList}
-              queuedQuestions={getQueueQuestions(queueIDList, mentorQuestions)}
-            />
-          </ListItem>
-        </List>
+          <RecordQueueBlock
+            classes={classes}
+            accessToken={props.accessToken}
+            recordQueue={recordQueue}
+          />
+        </ListItem>
       ) : undefined}
       <List
         data-cy="recording-blocks"
